scripts/slack/requirements.txt
flask
requests
python-dotenv
flask-sqlalchemy
psycopg2
